virtinstall: Make the --check path_exists= works
The path_exists validation doesn't work for virt-install because the
build_guest_instance() invokes the validate_disk() without passing in
the value of warn_overwrite which default is False, it causes the
check_path_exists() return too early so the path_definitely_exists()
doesn't have chance to be called.

After the patch, The --check path_exists takes effect.

Signed-off-by: Lin Ma <lma@suse.com>