virt-install: Add lifecycle event loop while installing a persistent domain
Currently this patch only handles the UNDEFINED lifecycle event.

We know that undefining a running persistent domain converts it to a
transient domain.
During installing a persistent domain, If the domain is undefined &
destroyed, The alive virt-install instance exits with an exception
because it tries to get info from a non-existent domain.

This patch introduces lifecycle event loop to let virt-install realize
the change of domain type (persistent -> transient), So in case domain
type changes to transient and is destroyed, The virt-install process can
correctly quit rather than try to get the domain info.

Resolves: https://github.com/virt-manager/virt-manager/issues/455

Signed-off-by: Lin Ma <lma@suse.com>