virt-install: add `OVMF_CODE_4M.fd` to UEFI arch patterns and fix catchall
Newer versions of OVMF package in Debian and Ubuntu has started to ship 4M
OVMF FW images with `_4M` suffix in their name, and updated the firmware
descriptor json files to point to them. The problem with that is, virt-install's
UEFI x86_64 arch pattern for OVMF does not explicitly cover the `_4M` suffix,
causing virt-install to fallback to capability based image selection and it
picks up an unintended firmware image on Debian and Ubuntu (i.e.,secure boot
enforced) as a result.

There is an existing catchall ".*OVMF.*" pattern that should've matched to the
new name, but it is not a raw string literal (lacks 'r' before dquote) so it
does not work as intended.

This patch adds the ".*OVMF_CODE_4M\.fd" pattern, and fixes the ".*OVMF.*"
catchall pattern.

Also, added an unit test to verify that `find_uefi_path_for_arch` behaves as
intended.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>