ci: checkout the code after installing git-core
This makes sure that the actions/checkout action uses git to fetch the
sources (and thus resulting in a git repository), rather than using the
REST API (and get a simple directory with sources).

This may be useful to certain services (e.g. newer version of
codecov-cli) as they need to have the sources as proper git repository.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>