ci: update github workflows to use meson
Running `meson dist` requires working git but that doesn't seem to be the
default in github actions when running jobs inside container. Any git
invocation fails with the following error:

fatal: detected dubious ownership in repository at '/__w/virt-manager/virt-manager'
To add an exception for this directory, call:

	git config --global --add safe.directory /__w/virt-manager/virt-manager

So that's what this patch does to make `meson dist` work correctly.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>