virt-clone: document limits of `--replace`
`--replace` will attempt to remove any existing VM with the new
clone name before performing the clone operation. There's a report
that it will fail for an existing VM with nvram, since we need
aren't doing the equivalent of `virsh undefine --nvram`

https://issues.redhat.com/browse/RHEL-65105

Making `--replace` remove nvram is easy, but it kinda breaks some
Cloner assumptions about determining ahead of time whether we
need to clone nvram or not. It's fixable but it's messy.

Also, making changes here raises the question about how we should be
handling all the other `virsh undefine` flags that have appeared over
the years.

I think it's better to leave `--replace` as is, and just document
that it's doing the simplest equivalent of
`virsh destroy` + `virsh undefine`. Anything more complicated
than that will cause the clone operation to fail, the user needs
to handle VM removal themselves.

Signed-off-by: Cole Robinson <crobinso@redhat.com>