storage: Add testsuite way to trigger StoragePool.install error
With the libvirt testdriver, we can create+start a directory
pool that points to any arbitrary host path. But this doesn't
match the behavior of the real storage driver, where pool start
will fail if the source directory doesn't exist.

Let's mock this by raising an exception if we are inside the
test suite and we see a special string in the pool XML

Signed-off-by: Cole Robinson <crobinso@redhat.com>