po: meson: drop --join-existing from gettext arguments
With setup.py we cleared po/virt-manager.pot file and created a new
fresh content using gettext.

I missed that fact when converting to meson and with the --join-existing
argument the comments for each msgid were growing with references to
their location in source files.

There is no need to clear the file and generate it again because
removing the --join-existing will end up with the same result.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>