Added ability to support Macvtap mode options
Added the ability for the user to configure the various modes of Macvtap when this option is selected. This allows for specific
configuration options, specifically allowing the user to configure mode private, VEPA, bridge, and passthrough. With the private
option all attached work loads are isolated but can communicate with external systems. Using external platforms/firewalls with 
ip-local-proxy-arp it is possible to facilitate micro-segmentation using this approach.