xmlapi: rewrite the code to use lxml instead of libxml2
The python module libxml2 is part of the libxml2 library and they are
planning to deprecate and remove this binding in favor of lxml that also
uses libxml2 but the binding is done in more pythonic way.

The change to use lxml uncovered some incorrectly formatted bits in our
test XMLs that are now fixed.

Disable new pylint category c-extension-no-member as lxml is not
introspected by default and pylint complains about it.

Resolves: https://github.com/virt-manager/virt-manager/issues/897
Signed-off-by: Pavel Hrdina <phrdina@redhat.com>