chore: avoid using 'l' as variable/parameter name
According to PEP8 [1], 'l' (lowercase 'L') is name to avoid, as
potentially hard to read.

Switch the occurrences to more clear names representing better what the
variables/parameters represent (e.g. "tunnel", "data", etc). In case the
variable/parameter is really a list, use "lst" to not use "list" and
override the builtin.

Additionally, a 'w' variable was changed for consistency with the 'l'
changed in the line above it.

There should be no behaviour changes.

[1] https://peps.python.org/pep-0008/#names-to-avoid

Signed-off-by: Pino Toscano <ptoscano@redhat.com>