virt-manager: improve feedback for unavailable system tray icon
In case the system tray icon is not available on Wayland (because of the
lack of Ayatana AppIndicator), the only notification of that is an
untranslated message shown only as tooltip on the check box of the
option.

To make that notification a bit more visible, use the same approach for
the libguestfs option change: add a row below the check box with a
warning icon and text next to it. This row is shown only in case there
is an error message to show, and it is used instead of setting the
tooltop of the check box.

Also, reword the error message to make it a bit less technical, and make
it translatable.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>