build: drop pylint from the test suite
pylint is more a static analyzer than a test suite, so it does not fit
too well when running the tests. In addition to that, it requires
additional Python modules needed only when running the UI tests:
https://github.com/virt-manager/virt-manager/issues/785

Since pylint now is run as github workflow for PRs, then drop the pylint
execution as part of the test suite.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>