tests: port pytest_ignore_collect() to pathlib
Create a pathlib version of the "path" argument of the
pytest_ignore_collect() hook; this will make it easier to the newer
version that uses pathlib directly.

This makes it possible to simplify the check for filenames to ignore:
since the filename is available, use it to do a quick lookup.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>